src/requirements.txt
asgiref==3.8.1
Django==4.2.17
image==1.5.33
ingescape==4.2.5
musicalbeeps==0.2.9
numpy==2.0.2
pillow==11.1.0
playsound==1.3.0
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-Quartz==10.3.2
simpleaudio==1.0.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
